+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/notify/BeanTransformerConfiguration.java
@@ -1,6 +1,5 @@
 package com.dwarfeng.familyhelper.webapi.node.configuration.notify;
 
-import com.dwarfeng.familyhelper.webapi.sdk.bean.notify.BeanMapper;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.notify.disp.JSFixedFastJsonDispMeta;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.notify.disp.JSFixedFastJsonDispNotifyHistory;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.notify.disp.JSFixedFastJsonDispNotifyInfoRecord;
@@ -22,82 +21,112 @@ public class BeanTransformerConfiguration {
     @Bean("notify.notifySettingBeanTransformer")
     public BeanTransformer<NotifySetting, JSFixedFastJsonNotifySetting> notifySettingBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                NotifySetting.class, JSFixedFastJsonNotifySetting.class, BeanMapper.class
+                NotifySetting.class, JSFixedFastJsonNotifySetting.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("notify.topicBeanTransformer")
     public BeanTransformer<Topic, FastJsonTopic> topicBeanTransformer() {
-        return new MapStructBeanTransformer<>(Topic.class, FastJsonTopic.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Topic.class, FastJsonTopic.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("notify.routerInfoBeanTransformer")
     public BeanTransformer<RouterInfo, JSFixedFastJsonRouterInfo> routerInfoBeanTransformer() {
-        return new MapStructBeanTransformer<>(RouterInfo.class, JSFixedFastJsonRouterInfo.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                RouterInfo.class, JSFixedFastJsonRouterInfo.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("notify.routerSupportBeanTransformer")
     public BeanTransformer<RouterSupport, FastJsonRouterSupport> routerSupportBeanTransformer() {
-        return new MapStructBeanTransformer<>(RouterSupport.class, FastJsonRouterSupport.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                RouterSupport.class, FastJsonRouterSupport.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("notify.dispatcherInfoBeanTransformer")
     public BeanTransformer<DispatcherInfo, FastJsonDispatcherInfo> dispatcherInfoBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispatcherInfo.class, FastJsonDispatcherInfo.class, BeanMapper.class
+                DispatcherInfo.class, FastJsonDispatcherInfo.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("notify.dispatcherSupportBeanTransformer")
     public BeanTransformer<DispatcherSupport, FastJsonDispatcherSupport> dispatcherSupportBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispatcherSupport.class, FastJsonDispatcherSupport.class, BeanMapper.class
+                DispatcherSupport.class, FastJsonDispatcherSupport.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("notify.senderInfoBeanTransformer")
     public BeanTransformer<SenderInfo, JSFixedFastJsonSenderInfo> senderInfoBeanTransformer() {
-        return new MapStructBeanTransformer<>(SenderInfo.class, JSFixedFastJsonSenderInfo.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                SenderInfo.class, JSFixedFastJsonSenderInfo.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("notify.senderSupportBeanTransformer")
     public BeanTransformer<SenderSupport, FastJsonSenderSupport> senderSupportBeanTransformer() {
-        return new MapStructBeanTransformer<>(SenderSupport.class, FastJsonSenderSupport.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                SenderSupport.class, FastJsonSenderSupport.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("notify.metaBeanTransformer")
     public BeanTransformer<Meta, JSFixedFastJsonMeta> metaBeanTransformer() {
-        return new MapStructBeanTransformer<>(Meta.class, JSFixedFastJsonMeta.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Meta.class, JSFixedFastJsonMeta.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("notify.dispMetaBeanTransformer")
     public BeanTransformer<DispMeta, JSFixedFastJsonDispMeta> dispMetaBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispMeta.class, JSFixedFastJsonDispMeta.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispMeta.class, JSFixedFastJsonDispMeta.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.notify.BeanMapper.class
+        );
     }
 
     @Bean("notify.metaIndicatorBeanTransformer")
     public BeanTransformer<MetaIndicator, FastJsonMetaIndicator> metaIndicatorBeanTransformer() {
-        return new MapStructBeanTransformer<>(MetaIndicator.class, FastJsonMetaIndicator.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                MetaIndicator.class, FastJsonMetaIndicator.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("notify.notifyHistoryBeanTransformer")
     public BeanTransformer<NotifyHistory, JSFixedFastJsonNotifyHistory> notifyHistoryBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                NotifyHistory.class, JSFixedFastJsonNotifyHistory.class, BeanMapper.class
+                NotifyHistory.class, JSFixedFastJsonNotifyHistory.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("notify.dispNotifyHistoryBeanTransformer")
     public BeanTransformer<DispNotifyHistory, JSFixedFastJsonDispNotifyHistory> dispNotifyHistoryBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispNotifyHistory.class, JSFixedFastJsonDispNotifyHistory.class, BeanMapper.class
+                DispNotifyHistory.class, JSFixedFastJsonDispNotifyHistory.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.notify.BeanMapper.class
         );
     }
 
     @Bean("notify.notifyInfoRecordBeanTransformer")
     public BeanTransformer<NotifyInfoRecord, JSFixedFastJsonNotifyInfoRecord> notifyInfoRecordBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                NotifyInfoRecord.class, JSFixedFastJsonNotifyInfoRecord.class, BeanMapper.class
+                NotifyInfoRecord.class, JSFixedFastJsonNotifyInfoRecord.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
         );
     }
 
@@ -105,14 +134,16 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<DispNotifyInfoRecord, JSFixedFastJsonDispNotifyInfoRecord>
     dispNotifyInfoRecordBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispNotifyInfoRecord.class, JSFixedFastJsonDispNotifyInfoRecord.class, BeanMapper.class
+                DispNotifyInfoRecord.class, JSFixedFastJsonDispNotifyInfoRecord.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.notify.BeanMapper.class
         );
     }
 
     @Bean("notify.notifySendRecordBeanTransformer")
     public BeanTransformer<NotifySendRecord, JSFixedFastJsonNotifySendRecord> notifySendRecordBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                NotifySendRecord.class, JSFixedFastJsonNotifySendRecord.class, BeanMapper.class
+                NotifySendRecord.class, JSFixedFastJsonNotifySendRecord.class,
+                com.dwarfeng.notify.sdk.bean.BeanMapper.class
         );
     }
 
@@ -120,7 +151,8 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<DispNotifySendRecord, JSFixedFastJsonDispNotifySendRecord>
     dispNotifySendRecordBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispNotifySendRecord.class, JSFixedFastJsonDispNotifySendRecord.class, BeanMapper.class
+                DispNotifySendRecord.class, JSFixedFastJsonDispNotifySendRecord.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.notify.BeanMapper.class
         );
     }
 }