+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/system/BeanTransformerConfiguration.java
@@ -2,7 +2,6 @@ package com.dwarfeng.familyhelper.webapi.node.configuration.system;
 
 import com.dwarfeng.acckeeper.sdk.bean.entity.*;
 import com.dwarfeng.acckeeper.stack.bean.entity.*;
-import com.dwarfeng.familyhelper.webapi.sdk.bean.system.BeanMapper;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.system.disp.FastJsonDispAccount;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.system.disp.FastJsonDispPermissionGroup;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.system.vo.FastJsonAccount;
@@ -27,59 +26,81 @@ public class BeanTransformerConfiguration {
 
     @Bean("system.accountBeanTransformer")
     public BeanTransformer<Account, FastJsonAccount> accountBeanTransformer() {
-        return new MapStructBeanTransformer<>(Account.class, FastJsonAccount.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Account.class, FastJsonAccount.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.system.BeanMapper.class
+        );
     }
 
     @Bean("system.dispAccountBeanTransformer")
     public BeanTransformer<DispAccount, FastJsonDispAccount> dispAccountBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispAccount.class, FastJsonDispAccount.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispAccount.class, FastJsonDispAccount.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.system.BeanMapper.class
+        );
     }
 
     @Bean("system.permissionBeanTransformer")
     public BeanTransformer<Permission, FastJsonPermission> permissionBeanTransformer() {
-        return new MapStructBeanTransformer<>(Permission.class, FastJsonPermission.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Permission.class, FastJsonPermission.class,
+                com.dwarfeng.rbacds.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("system.roleBeanTransformer")
     public BeanTransformer<Role, FastJsonRole> roleBeanTransformer() {
-        return new MapStructBeanTransformer<>(Role.class, FastJsonRole.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Role.class, FastJsonRole.class,
+                com.dwarfeng.rbacds.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("system.pexpBeanTransformer")
     public BeanTransformer<Pexp, JSFixedFastJsonPexp> pexpBeanTransformer() {
-        return new MapStructBeanTransformer<>(Pexp.class, JSFixedFastJsonPexp.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Pexp.class, JSFixedFastJsonPexp.class,
+                com.dwarfeng.rbacds.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("system.permissionGroupBeanTransformer")
     public BeanTransformer<PermissionGroup, FastJsonPermissionGroup> permissionGroupBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                PermissionGroup.class, FastJsonPermissionGroup.class, BeanMapper.class
+                PermissionGroup.class, FastJsonPermissionGroup.class,
+                com.dwarfeng.rbacds.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("system.dispPermissionGroupBeanTransformer")
     public BeanTransformer<DispPermissionGroup, FastJsonDispPermissionGroup> dispPermissionGroupBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispPermissionGroup.class, FastJsonDispPermissionGroup.class, BeanMapper.class
+                DispPermissionGroup.class, FastJsonDispPermissionGroup.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.system.BeanMapper.class
         );
     }
 
     @Bean("system.loginStateBeanTransformer")
     public BeanTransformer<LoginState, JSFixedFastJsonLoginState> loginStateBeanTransformer() {
-        return new MapStructBeanTransformer<>(LoginState.class, JSFixedFastJsonLoginState.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                LoginState.class, JSFixedFastJsonLoginState.class,
+                com.dwarfeng.acckeeper.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("system.loginHistoryBeanTransformer")
     public BeanTransformer<LoginHistory, JSFixedFastJsonLoginHistory> loginHistoryBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                LoginHistory.class, JSFixedFastJsonLoginHistory.class, BeanMapper.class
+                LoginHistory.class, JSFixedFastJsonLoginHistory.class,
+                com.dwarfeng.acckeeper.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("system.loginParamRecordBeanTransformer")
     public BeanTransformer<LoginParamRecord, JSFixedFastJsonLoginParamRecord> loginParamRecordBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                LoginParamRecord.class, JSFixedFastJsonLoginParamRecord.class, BeanMapper.class
+                LoginParamRecord.class, JSFixedFastJsonLoginParamRecord.class,
+                com.dwarfeng.acckeeper.sdk.bean.BeanMapper.class
         );
     }
 
@@ -87,14 +108,16 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<ProtectDetailRecord, JSFixedFastJsonProtectDetailRecord>
     protectDetailRecordBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ProtectDetailRecord.class, JSFixedFastJsonProtectDetailRecord.class, BeanMapper.class
+                ProtectDetailRecord.class, JSFixedFastJsonProtectDetailRecord.class,
+                com.dwarfeng.acckeeper.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("system.deriveHistoryBeanTransformer")
     public BeanTransformer<DeriveHistory, JSFixedFastJsonDeriveHistory> deriveHistoryBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DeriveHistory.class, JSFixedFastJsonDeriveHistory.class, BeanMapper.class
+                DeriveHistory.class, JSFixedFastJsonDeriveHistory.class,
+                com.dwarfeng.acckeeper.sdk.bean.BeanMapper.class
         );
     }
 }