+++ familyhelper-webapi-sdk/src/main/java/com/dwarfeng/familyhelper/webapi/sdk/bean/settingrepo/BeanMapper.java
@@ -1,8 +1,5 @@
 package com.dwarfeng.familyhelper.webapi.sdk.bean.settingrepo;
 
-import com.dwarfeng.settingrepo.sdk.bean.entity.*;
-import com.dwarfeng.settingrepo.stack.bean.entity.*;
-import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 
 /**
@@ -16,34 +13,4 @@ import org.mapstruct.Mapper;
  */
 @Mapper
 public interface BeanMapper {
-
-    FastJsonSettingNode settingNodeToFastJson(SettingNode settingNode);
-
-    @InheritInverseConfiguration
-    SettingNode settingNodeFromFastJson(FastJsonSettingNode fastJsonSettingNode);
-
-    FastJsonSettingCategory settingCategoryToFastJson(SettingCategory settingCategory);
-
-    @InheritInverseConfiguration
-    SettingCategory settingCategoryFromFastJson(FastJsonSettingCategory fastJsonSettingCategory);
-
-    FastJsonFormatterSupport formatterSupportToFastJson(FormatterSupport formatterSupport);
-
-    @InheritInverseConfiguration
-    FormatterSupport formatterSupportFromFastJson(FastJsonFormatterSupport fastJsonFormatterSupport);
-
-    FastJsonTextNode textNodeToFastJson(TextNode textNode);
-
-    @InheritInverseConfiguration
-    TextNode textNodeFromFastJson(FastJsonTextNode fastJsonTextNode);
-
-    FastJsonImageNode imageNodeToFastJson(ImageNode imageNode);
-
-    @InheritInverseConfiguration
-    ImageNode imageNodeFromFastJson(FastJsonImageNode fastJsonImageNode);
-
-    FastJsonImageListNode imageListNodeToFastJson(ImageListNode imageListNode);
-
-    @InheritInverseConfiguration
-    ImageListNode imageListNodeFromFastJson(FastJsonImageListNode fastJsonImageListNode);
 }