+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 #### 功能构建
 
+- 优化实体映射器机制。
+  - com.dwarfeng.familyhelper.webapi.sdk.bean.notify.BeanMapper。
+  - com.dwarfeng.familyhelper.webapi.sdk.bean.settingrepo.BeanMapper。
+  - com.dwarfeng.familyhelper.webapi.sdk.bean.system.BeanMapper。
+
 - 优化项目结构。
   - 将 `node` 模块中的 `configuration` 包中的 `Mapper` 接口提升至 `sdk` 模块的 `bean` 包中，并优化接口名称。
   - 将 `stack` 模块中的 `bean` 包优化为 `bean.submodule.category` 模式。